+++ addOns/authhelper/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update dependency.
 - Include the Zest statement index in the authentication diagnostics' steps.
+- Include Origin header when doing authentication verification.
 
 ## [0.40.0] - 2026-06-12
 ### Added