+++ addOns/authhelper/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [0.41.0] - 2026-08-07
 ### Added
 - Automation Framework `diagnostics` job to start and stop plan-level recording of authentication related diagnostics.
 
@@ -389,6 +389,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Support of authentication request identification and configuration.
 
+[0.41.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.41.0
 [0.40.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.40.0
 [0.39.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.39.0
 [0.38.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.38.0